Revert "- make add template screen design"
This reverts commit 0aa371b2eea3b26219b2bf6b6bdf20316522732f.
