--- flutter.md
@@ -1,94 +1,0 @@
-# Internship Program Overview (10 Weeks)
-
-## Training (4 weeks)
-
-### Week 1: Orientation and Soft Skills
-
-- *Introduction to the Company:*
-  - Overview of the company’s mission, values, and culture.
-  - Introduction to team members and key contacts.
-- *Soft Skills Workshops:*
-  - Communication skills (verbal and written).
-  - Time management and productivity techniques.
-  - Professional etiquette and workplace behavior.
-- *Team-Building Activities:*
-  - Ice-breaking sessions.
-  - Collaborative tasks to build rapport among interns.
-
-### Week 2: Flutter Basics
-
-- *Introduction to Flutter:*
-  - Overview of Flutter framework and Dart language.
-  - Setting up the development environment.
-  - Understanding the Flutter architecture.
-- *Dart Programming Fundamentals:*
-  - Core concepts: variables, data types, control structures, functions.
-  - Object-oriented programming: classes, objects, inheritance.
-  - Asynchronous programming: futures, async/await.
-- *Basic Flutter Widgets:*
-  - StatelessWidget and StatefulWidget.
-  - Basic widgets: Text, Container, Column, Row, ListView.
-  - Handling user input: TextField, Buttons, GestureDetector.
-
-### Week 3: Intermediate Flutter Development
-
-- *State Management:*
-  - Introduction to state management concepts.
-  - setState, InheritedWidget.
-  - Provider package for state management.
-- *Navigation and Routing:*
-  - Navigation basics: Navigator, routes.
-  - Passing data between screens.
-  - Named routes and navigation with arguments.
-- *Working with APIs:*
-  - Making HTTP requests using the `http` package.
-  - Parsing JSON data.
-  - Displaying data from an API in the UI.
-- *Advanced UI Components:*
-  - Custom widgets and composition.
-  - Animations: implicit and explicit animations.
-  - Responsive design: adapting to different screen sizes.
-
-### Week 4: Training on Documenting and Team Skills
-
-- *Documentation Best Practices:*
-  - Writing clear and concise technical documentation.
-  - Documenting code with comments and external documentation tools.
-- *Collaboration Tools and Techniques:*
-  - Version control with Git: branching, merging, pull requests.
-  - Project management tools: Jira, Trello.
-  - Agile methodologies and sprint planning.
-- *Team Collaboration Exercises:*
-  - Working in small groups on mini-projects.
-  - Peer code reviews and feedback sessions.
-
-## Technical Training and Project (6 weeks)
-
-### Week 5: Project Introduction and Preparation
-
-- *Project Overview:*
-  - Explanation of project goals and objectives.
-  - Initial project planning and brainstorming.
-  - Setting up the project environment and tools.
-- *Technology Introduction:*
-  - Overview of technologies and tools to be used in the project.
-  - Initial coding exercises related to the project.
-
-### Weeks 6-9: Development Period
-
-- *General Development Work:*
-  - Implementing project features.
-  - Regular progress check-ins and reviews.
-
-### Week 10: Finalization and Deployment
-
-- *Final Review and Polishing:*
-  - Code reviews and optimization.
-  - Finalizing documentation and project deliverables.
-- *Deployment:*
-  - Preparing the project for deployment.
-  - Deployment to a staging or production environment.
-- *Presentation:*
-  - Preparing a presentation to showcase the project.
-  - Presenting the project to the team and stakeholders.
-  - Gathering feedback and reflecting on the internship experience.
