+++ lib/core/networking/dio_factory.dart
@@ -46,7 +46,6 @@ class DioFactory {
         requestBody: true,
         requestHeader: true,
         responseHeader: true,
-        responseBody: true,
       ),
     );
   }
