+++ lib/core/routing/routes.dart
@@ -5,6 +5,5 @@ class Routes {
   static const String homeScreen = '/homeScreen';
   static const String manageAudiances = '/manageAudiances';
   static const String templatesScreen = '/templatesScreen';
-  static const String addTemplateScreen = '/addTemplateScreen';
 
 }
