+++ lib/core/theming/colors.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
 
 class ColorsManager {
-  static const Color darkBackGround = Color(0xFF0F172A);
-  static const Color darkAppBarBackGround = Color(0xFF1E293B);
+  static const Color darkBackGround = Color(0xFF1e293b);
+  static const Color darkAppBarBackGround = Color(0xFF334155);
   static const Color containerTitleColor = Color(0xFF84cc16);
   static const Color saerchTextFieldColor = Color(0xFF334155);
-  static const Color saerchTextFieldBackGroundColor = Color(0xFF334155);
-  static const Color saerchTextFieldHintColor = Color(0xFF94A3B8);
-  static const Color limeColor = Color(0xFF84CC16);
-  static const Color buttonColor = Color(0xFFEAB308);
-  static Color? teal400 = const Color(0xff2DD4BF);
+    static const Color saerchTextFieldBackGroundColor = Color(0xFF334155);
+    static const Color saerchTextFieldHintColor = Color(0xFF94A3B8);
+
 
   // static const Color darkBlue = Color(0xFF242424);
   // static const Color gray = Color(0xFF757575);
