+++ lib/features/manage_audiances/presentation/pages/manage_audiances_screen.dart
@@ -1,5 +1,5 @@
 import 'package:bulk_app/core/resources/app_strings.dart';
-import 'package:bulk_app/features/manage_audiances/presentation/widgets/manage_audiances_widgets/manage_audiances_body.dart';
+import 'package:bulk_app/features/manage_audiances/presentation/widgets/manage_audiances_body.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
