--- lib/features/templates/presentation/cubit/add_template_cubit/add_template_state.dart
@@ -1,6 +1,0 @@
-part of 'add_template_cubit.dart';
-
-@freezed
-class AddTemplateState with _$AddTemplateState {
-  const factory AddTemplateState.initial() = _Initial;
-}
