+++ lib/features/templates/presentation/cubit/cubit/templates_cubit.dart
