+++ lib/features/templates/presentation/cubit/cubit/templates_state.dart
