+++ lib/features/templates/presentation/pages/templates_screen.dart
@@ -2,7 +2,7 @@ import 'package:bulk_app/core/resources/app_strings.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import '../../../../core/widgets/custom_app_bar.dart';
-import '../widgets/templates_screen_widgets/templates_screen_body.dart';
+import '../widgets/templates_screen_body.dart';
 
 class TemplatesScreen extends StatelessWidget {
   const TemplatesScreen({super.key});
