+++ lib/features/templates/presentation/widgets/templates_screen_body.dart
@@ -3,9 +3,8 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../../../core/resources/app_strings.dart';
-import '../../../../../core/theming/colors.dart';
-import '../../../../../core/widgets/app_search_field.dart';
+import '../../../../core/resources/app_strings.dart';
+import '../../../../core/widgets/app_search_field.dart';
 import 'templates_list_view.dart';
 
 class TemplatesBody extends StatelessWidget {
@@ -17,13 +16,8 @@ class TemplatesBody extends StatelessWidget {
       padding: EdgeInsets.symmetric(horizontal: 20.w, vertical: 20.h),
       child: Column(
         children: [
-          AppTextField(
+          AppSearchField(
             hintText: AppStrings.searchTemplates.tr(),
-            prefixIcon: Icon(
-              Icons.search,
-              size: 25.sp,
-              color: ColorsManager.saerchTextFieldHintColor,
-            ),
           ),
           20.sizedBoxHeight,
           const Expanded(child: TemplatesListView()),
