+++ pubspec.yaml
@@ -63,6 +63,5 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
-    - assets/icons/
     - assets/translation/en-US.json
     - assets/translation/ar-SA.json
