--- traini.md
@@ -1,99 +1,0 @@
-# Internship Program Overview (10 Weeks)
-
-## Training (4 weeks)
-
-### Week 1: Orientation and Soft Skills
-
-- *Introduction to the Company:*
-  - Overview of the company’s mission, values, and culture.
-  - Introduction to team members and key contacts.
-- *Soft Skills Workshops:*
-  - Communication skills (verbal and written).
-  - Time management and productivity techniques.
-  - Professional etiquette and workplace behavior.
-- *Team-Building Activities:*
-  - Ice-breaking sessions.
-  - Collaborative tasks to build rapport among interns.
-
-### Week 2: Basic Technology Training
-
-- *Introduction to Flutter:*
-  - Overview of Flutter framework and Dart language.
-  - Setting up the development environment.
-  - Understanding the Flutter architecture.
-  
-- *Dart Programming Fundamentals:*
-  - Core concepts: variables, data types, control structures, functions.
-  - Object-oriented programming: classes, objects, inheritance.
-  - Asynchronous programming: futures, async/await.
-
-- *Basic Flutter Widgets:*
-  - StatelessWidget and StatefulWidget.
-  - Basic widgets: Text, Container, Column, Row, ListView.
-  - Handling user input: TextField, Buttons, GestureDetector.
-
-### Week 3: Intermediate Technology Training
-
-- *State Management:*
-  - Introduction to state management concepts.
-  - Provider,Bloc
-  
-- *Working with APIs:*
-  - Making HTTP requests using the http and dio packages.
-  - Parsing JSON data.
-  - Displaying data from an API in the UI.
-
-- *Caching Data:*
-  - Caching data using the shared_preferences package.
-  - Caching data using the hive package.
-
-- *Advanced UI Components:*
-  - Custom widgets and composition.
-  - Responsive design: adapting to different screen sizes.
-  
-- *Clean Architecture:*
-  - Introduction to clean architecture concepts and patterns.
-
-### Week 4: Training on Documenting and Team Skills
-
-- *Documentation Best Practices:*
-  - Writing clear and concise technical documentation.
-  - Documenting code with comments and external documentation tools.
-- *Collaboration Tools and Techniques:*
-  - Version control with Git: branching, merging, pull requests.
-  - Project management tools: Jira, Trello.
-  - Agile methodologies and sprint planning.
-- *Team Collaboration Exercises:*
-  - Working in small groups on mini-projects.
-  - Peer code reviews and feedback sessions.
-
-## Technical Training and Project (6 weeks)
-
-### Week 5: Project Introduction and Preparation
-
-- *Project Overview:*
-  - Explanation of project goals and objectives.
-  - Initial project planning and brainstorming.
-  - Setting up the project environment and tools.
-- *Technology Introduction:*
-  - Overview of technologies and tools to be used in the project.
-  - Initial coding exercises related to the project.
-
-### Weeks 6-9: Development Period
-
-- *General development work:*
-  - Implementing project features.
-  - Regular progress check-ins and reviews.
-
-### Week 10: Finalization and Deployment
-
-- *Final Review and Polishing:*
-  - Code reviews and optimization.
-  - Finalizing documentation and project deliverables.
-- *Deployment:*
-  - Preparing the project for deployment.
-  - Deployment to a staging or production environment.
-- *Presentation:*
-  - Preparing a presentation to showcase the project.
-  - Presenting the project to the team and stakeholders.
-  - Gathering feedback and reflecting on the internship experience.
